Complete Overhaul
Decided to rework the former GameState class into a Game class which spreads out into many other classes, the GameState class was just too much of a hassle to deal with as it was.
Added new resources and a Menu to the game.
Added a messy mouse listener and interface for it.
Messed around with Enums to dictate bead colour and gamestate.